requirements.txt
# Core
ultralytics>=8.0.0
opencv-python>=4.8.0
streamlit>=1.28.0
plotly>=5.18.0
click>=8.1.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.11.0
imageio>=2.31.0
python-dotenv>=1.0
shapely>=2.0.0
pandas>=2.0.0

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
